Add RGB-to-Label Script
Turn an RGB image into labels.  Solid red is mapped to 0x01, solid black
is mapped to 0x02, and everything else is mapped to 0x00.  The idea is
that one can use GIMP or photoshop or similar to label an RGB image,
then use this script to turn that into a single band label file.